By default, don't read stdin when using reserve as a module.
I'm specifically doing this to avoid a situation where a reserve-using
server is run as a systemd job and, because systemd connects its stdin
to /dev/null, the job immediately exits. I think it's surprising for a
random web server which happens to use reserve.FileServer to be
sensitive to EOF or even to read stdin at all, so I'm making it opt-in.
Maybe having separate flags for reading stdin and exiting on EOF would
make sense, too.